dom/tests/mochitest/general/test_performance_nav_timing_before_onload.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1405961 -  Using PerformanceNavigationTiming before onload</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <div id="content"> </div>
  <script>
  SimpleTest.waitForExplicitFinish();

  SpecialPowers.pushPrefEnv({"set": [["dom.enable_performance_navigation_timing", true]]}, start);

  function start() {
    var p = performance.getEntriesByName(window.location.href)[0];
    ok(!!p, "There should be an entry for the document");
    document.getElementById("content").textContent += JSON.stringify(p);

    SimpleTest.finish();
  }
  </script>
</body>
</html>
